Revert "[SPARK-24355] Spark external shuffle server improvement to better handle block fetch requests. (#89)"
This reverts commit a0d72e8245d5b4227f12a8ef1652b8faa92fcec3.
